code/requirements.txt
pyyaml
pytorch==1.4
scipy
opencv
tensorboardX
scikit-learn
easydict
imgaug
dxfgrabber